requirements.txt
streamlit>=1.39.0
easyocr>=1.7.2
Pillow>=10.3.0
numpy>=1.26.0
groq>=0.8.0
transformers>=4.46.0
torch>=2.4.0
python-dotenv>=1.0.1
pip install hf_xet